style(logging): move setloglvl as prestep in rootcmd
